Renames all actions attached to the HTTP method `GET`.
