+++ src/Actions/Api/Get/CurrentTrackAction.php
@@ -9,17 +9,19 @@ use CodeKandis\Tiphy\Persistence\MariaDb\Connector;
 use CodeKandis\Tiphy\Persistence\MariaDb\ConnectorInterface;
 use CodeKandis\Tiphy\Persistence\PersistenceException;
 use CodeKandis\TradioApi\Configurations\ConfigurationRegistry;
+use CodeKandis\TradioApi\Entities\CurrentTrackEntity;
 use CodeKandis\TradioApi\Entities\FavoriteEntity;
-use CodeKandis\TradioApi\Entities\UriExtenders\FavoriteUriExtender;
-use CodeKandis\TradioApi\Entities\UserEntity;
-use CodeKandis\TradioApi\Errors\UsersErrorCodes;
-use CodeKandis\TradioApi\Errors\UsersErrorMessages;
+use CodeKandis\TradioApi\Entities\StationEntity;
+use CodeKandis\TradioApi\Entities\UriExtenders\CurrentTrackUriExtender;
+use CodeKandis\TradioApi\Errors\StationsErrorCodes;
+use CodeKandis\TradioApi\Errors\StationsErrorMessages;
+use CodeKandis\TradioApi\Http\Readers\CurrentTrackReader;
 use CodeKandis\TradioApi\Http\UriBuilders\ApiUriBuilder;
 use CodeKandis\TradioApi\Persistence\MariaDb\Repositories\FavoritesRepository;
-use CodeKandis\TradioApi\Persistence\MariaDb\Repositories\UsersRepository;
+use CodeKandis\TradioApi\Persistence\MariaDb\Repositories\StationsRepository;
 use ReflectionException;
 
-class GetUserFavoritesAction extends AbstractAction
+class CurrentTrackAction extends AbstractAction
 {
 	/** @var ConnectorInterface */
 	private $databaseConnector;
@@ -57,25 +59,29 @@ class GetUserFavoritesAction extends AbstractAction
 	{
 		$inputData = $this->getInputData();
 
-		$requestedUser     = new UserEntity();
-		$requestedUser->id = $inputData[ 'id' ];
-		$user              = $this->readUser( $requestedUser );
+		$requestedStation     = new StationEntity();
+		$requestedStation->id = $inputData[ 'id' ];
+		$station              = $this->readStation( $requestedStation );
 
-		if ( null === $user )
+		if ( null === $station )
 		{
-			$errorInformation = new ErrorInformation( UsersErrorCodes::USER_UNKNOWN, UsersErrorMessages::USER_UNKNOWN, $inputData );
+			$errorInformation = new ErrorInformation( StationsErrorCodes::STATION_UNKNOWN, StationsErrorMessages::STATION_UNKNOWN, $inputData );
 			( new JsonResponder( StatusCodes::NOT_FOUND, null, $errorInformation ) )
 				->respond();
 
 			return;
 		}
 
-		$favorites = $this->readUsersFavorites( $user );
-		$this->extendUris( $favorites );
+		$currentTrack            = $this->readCurrentTrack( $station );
+		$requestedFavorite       = new FavoriteEntity();
+		$requestedFavorite->name = $currentTrack->name;
+		$favorite                = $this->readFavoriteByTrackName( $requestedFavorite );
+		$this->extendUris( $currentTrack, $station, $favorite );
 
 		$responderData = [
-			'favorites' => $favorites,
+			'currentTrack' => $currentTrack
 		];
+
 		( new JsonResponder( StatusCodes::OK, $responderData ) )
 			->respond();
 	}
@@ -88,39 +94,42 @@ class GetUserFavoritesAction extends AbstractAction
 		return $this->arguments;
 	}
 
-	/**
-	 * @param FavoriteEntity[] $favorites
-	 */
-	private function extendUris( array $favorites ): void
+	private function extendUris( CurrentTrackEntity $currentTrack, StationEntity $station, ?FavoriteEntity $favorite ): void
 	{
 		$uriBuilder = $this->getUriBuilder();
-		foreach ( $favorites as $favorite )
-		{
-			( new FavoriteUriExtender( $uriBuilder, $favorite ) )
-				->extend();
-		}
+		( new CurrentTrackUriExtender( $uriBuilder, $currentTrack, $station, $favorite ) )
+			->extend();
 	}
 
 	/**
 	 * @throws PersistenceException
 	 */
-	private function readUser( UserEntity $requestedUser ): ?UserEntity
+	private function readStation( StationEntity $requestedStation ): ?StationEntity
 	{
 		$databaseConnector = $this->getDatabaseConnector();
 
-		return ( new UsersRepository( $databaseConnector ) )
-			->readUserById( $requestedUser );
+		return ( new StationsRepository( $databaseConnector ) )
+			->readStationById( $requestedStation );
+	}
+
+	private function readCurrentTrack( StationEntity $station ): CurrentTrackEntity
+	{
+		$currentTrack            = new CurrentTrackEntity();
+		$currentTrack->name      = ( new CurrentTrackReader() )
+			->read( $station->tracklistUri, $station->currentTrackXPath );
+		$currentTrack->stationId = $station->id;
+
+		return $currentTrack;
 	}
 
 	/**
-	 * @return FavoriteEntity[]
 	 * @throws PersistenceException
 	 */
-	private function readUsersFavorites( UserEntity $user ): array
+	private function readFavoriteByTrackName( FavoriteEntity $requestedFavorite ): ?FavoriteEntity
 	{
 		$databaseConnector = $this->getDatabaseConnector();
 
 		return ( new FavoritesRepository( $databaseConnector ) )
-			->readFavoritesByUserId( $user );
+			->readFavoriteByTrackName( $requestedFavorite );
 	}
 }
