+++ src/Actions/Api/Get/FavoriteStationsAction.php
@@ -9,17 +9,17 @@ use CodeKandis\Tiphy\Persistence\MariaDb\Connector;
 use CodeKandis\Tiphy\Persistence\MariaDb\ConnectorInterface;
 use CodeKandis\Tiphy\Persistence\PersistenceException;
 use CodeKandis\TradioApi\Configurations\ConfigurationRegistry;
+use CodeKandis\TradioApi\Entities\FavoriteEntity;
 use CodeKandis\TradioApi\Entities\StationEntity;
 use CodeKandis\TradioApi\Entities\UriExtenders\StationUriExtender;
-use CodeKandis\TradioApi\Entities\UserEntity;
-use CodeKandis\TradioApi\Errors\UsersErrorCodes;
-use CodeKandis\TradioApi\Errors\UsersErrorMessages;
+use CodeKandis\TradioApi\Errors\FavoritesErrorCodes;
+use CodeKandis\TradioApi\Errors\FavoritesErrorMessages;
 use CodeKandis\TradioApi\Http\UriBuilders\ApiUriBuilder;
+use CodeKandis\TradioApi\Persistence\MariaDb\Repositories\FavoritesRepository;
 use CodeKandis\TradioApi\Persistence\MariaDb\Repositories\StationsRepository;
-use CodeKandis\TradioApi\Persistence\MariaDb\Repositories\UsersRepository;
 use ReflectionException;
 
-class GetUserStationsAction extends AbstractAction
+class FavoriteStationsAction extends AbstractAction
 {
 	/** @var ConnectorInterface */
 	private $databaseConnector;
@@ -57,20 +57,20 @@ class GetUserStationsAction extends AbstractAction
 	{
 		$inputData = $this->getInputData();
 
-		$requestedUser     = new UserEntity();
-		$requestedUser->id = $inputData[ 'id' ];
-		$user              = $this->readUser( $requestedUser );
+		$requestedFavorite     = new FavoriteEntity();
+		$requestedFavorite->id = $inputData[ 'id' ];
+		$favorite              = $this->readFavorite( $requestedFavorite );
 
-		if ( null === $user )
+		if ( null === $favorite )
 		{
-			$errorInformation = new ErrorInformation( UsersErrorCodes::USER_UNKNOWN, UsersErrorMessages::USER_UNKNOWN, $inputData );
+			$errorInformation = new ErrorInformation( FavoritesErrorCodes::FAVORITE_UNKNOWN, FavoritesErrorMessages::FAVORITE_UNKNOWN, $inputData );
 			( new JsonResponder( StatusCodes::NOT_FOUND, null, $errorInformation ) )
 				->respond();
 
 			return;
 		}
 
-		$stations = $this->readUsersStations( $user );
+		$stations = $this->readFavoriteStations( $favorite );
 		$this->extendUris( $stations );
 
 		$responderData = [
@@ -104,23 +104,23 @@ class GetUserStationsAction extends AbstractAction
 	/**
 	 * @throws PersistenceException
 	 */
-	private function readUser( UserEntity $requestedUser ): ?UserEntity
+	private function readFavorite( FavoriteEntity $favorite ): ?FavoriteEntity
 	{
 		$databaseConnector = $this->getDatabaseConnector();
 
-		return ( new UsersRepository( $databaseConnector ) )
-			->readUserById( $requestedUser );
+		return ( new FavoritesRepository( $databaseConnector ) )
+			->readFavoriteById( $favorite );
 	}
 
 	/**
 	 * @return StationEntity[]
 	 * @throws PersistenceException
 	 */
-	private function readUsersStations( UserEntity $user ): array
+	private function readFavoriteStations( FavoriteEntity $favorite ): array
 	{
 		$databaseConnector = $this->getDatabaseConnector();
 
 		return ( new StationsRepository( $databaseConnector ) )
-			->readStationsByUserId( $user );
+			->readStationsByFavoriteId( $favorite );
 	}
 }
