+++ src/Actions/Api/Get/FavoriteUsersAction.php
@@ -1,0 +1,126 @@
+<?php declare( strict_types = 1 );
+namespace CodeKandis\TradioApi\Actions\Api\Get;
+
+use CodeKandis\Tiphy\Actions\AbstractAction;
+use CodeKandis\Tiphy\Exceptions\ErrorInformation;
+use CodeKandis\Tiphy\Http\Responses\JsonResponder;
+use CodeKandis\Tiphy\Http\Responses\StatusCodes;
+use CodeKandis\Tiphy\Persistence\MariaDb\Connector;
+use CodeKandis\Tiphy\Persistence\MariaDb\ConnectorInterface;
+use CodeKandis\Tiphy\Persistence\PersistenceException;
+use CodeKandis\TradioApi\Configurations\ConfigurationRegistry;
+use CodeKandis\TradioApi\Entities\FavoriteEntity;
+use CodeKandis\TradioApi\Entities\UriExtenders\UserUriExtender;
+use CodeKandis\TradioApi\Entities\UserEntity;
+use CodeKandis\TradioApi\Errors\FavoritesErrorCodes;
+use CodeKandis\TradioApi\Errors\FavoritesErrorMessages;
+use CodeKandis\TradioApi\Http\UriBuilders\ApiUriBuilder;
+use CodeKandis\TradioApi\Persistence\MariaDb\Repositories\FavoritesRepository;
+use CodeKandis\TradioApi\Persistence\MariaDb\Repositories\UsersRepository;
+use ReflectionException;
+
+class FavoriteUsersAction extends AbstractAction
+{
+	/** @var ConnectorInterface */
+	private $databaseConnector;
+
+	/** @var ApiUriBuilder */
+	private $uriBuilder;
+
+	private function getDatabaseConnector(): ConnectorInterface
+	{
+		if ( null === $this->databaseConnector )
+		{
+			$databaseConfig          = ConfigurationRegistry::_()->getPersistenceConfiguration();
+			$this->databaseConnector = new Connector( $databaseConfig );
+		}
+
+		return $this->databaseConnector;
+	}
+
+	private function getUriBuilder(): ApiUriBuilder
+	{
+		if ( null === $this->uriBuilder )
+		{
+			$uriBuilderConfiguration = ConfigurationRegistry::_()->getUriBuilderConfiguration();
+			$this->uriBuilder        = new ApiUriBuilder( $uriBuilderConfiguration );
+		}
+
+		return $this->uriBuilder;
+	}
+
+	/**
+	 * @throws PersistenceException
+	 * @throws ReflectionException
+	 */
+	public function execute(): void
+	{
+		$inputData = $this->getInputData();
+
+		$requestedFavorite     = new FavoriteEntity();
+		$requestedFavorite->id = $inputData[ 'id' ];
+		$favorite              = $this->readFavorite( $requestedFavorite );
+
+		if ( null === $favorite )
+		{
+			$errorInformation = new ErrorInformation( FavoritesErrorCodes::FAVORITE_UNKNOWN, FavoritesErrorMessages::FAVORITE_UNKNOWN, $inputData );
+			( new JsonResponder( StatusCodes::NOT_FOUND, null, $errorInformation ) )
+				->respond();
+
+			return;
+		}
+
+		$users = $this->readFavoriteUsers( $favorite );
+		$this->extendUris( $users );
+
+		$responderData = [
+			'users' => $users,
+		];
+		( new JsonResponder( StatusCodes::OK, $responderData ) )
+			->respond();
+	}
+
+	/**
+	 * @return string[]
+	 */
+	private function getInputData(): array
+	{
+		return $this->arguments;
+	}
+
+	/**
+	 * @param UserEntity[] $users
+	 */
+	private function extendUris( array $users ): void
+	{
+		$uriBuilder = $this->getUriBuilder();
+		foreach ( $users as $user )
+		{
+			( new UserUriExtender( $uriBuilder, $user ) )
+				->extend();
+		}
+	}
+
+	/**
+	 * @throws PersistenceException
+	 */
+	private function readFavorite( FavoriteEntity $favorite ): ?FavoriteEntity
+	{
+		$databaseConnector = $this->getDatabaseConnector();
+
+		return ( new FavoritesRepository( $databaseConnector ) )
+			->readFavoriteById( $favorite );
+	}
+
+	/**
+	 * @return UserEntity[]
+	 * @throws PersistenceException
+	 */
+	private function readFavoriteUsers( FavoriteEntity $favorite ): array
+	{
+		$databaseConnector = $this->getDatabaseConnector();
+
+		return ( new UsersRepository( $databaseConnector ) )
+			->readUsersByFavoriteId( $favorite );
+	}
+}
